[faster syncing through limiting peers] moved syncing % math all to a single function

+++ src/rpcserver.h
@@ -291,7 +291,7 @@ extern UniValue getpeerinfo(const UniValue& params, bool fHelp);
 /*//////////////////////////////peer syncing speed//////////////////////////////*/
 extern void limit_run();
 extern void limit_peers(int i);
-extern double getVerificationProgress_RPC(const CBlockIndex *tipIn);
+extern int getVerificationProgress_RPC(const CBlockIndex *tipIn);
 /*//////////////////////////////peer syncing speed//////////////////////////////*/
 extern UniValue ping(const UniValue& params, bool fHelp);
 extern UniValue addnode(const UniValue& params, bool fHelp);
